refactor: extract navigation types
