+++ app/[locale]/_components/app-header.tsx
@@ -1,14 +1,11 @@
 import { useTranslations } from "next-intl";
 import type { ReactNode } from "react";
 
-import {
-	AppNavigation,
-	AppNavigationMobile,
-	type NavigationItem,
-} from "@/app/[locale]/_components/app-navigation";
+import { AppNavigation, AppNavigationMobile } from "@/app/[locale]/_components/app-navigation";
 import { ColorSchemeSwitcher } from "@/app/[locale]/_components/color-scheme-switcher";
 import { LocaleSwitcher } from "@/app/[locale]/_components/locale-switcher";
 import { createHref } from "@/lib/navigation/create-href";
+import type { NavigationItem } from "@/lib/navigation/navigation";
 
 export function AppHeader(): ReactNode {
 	const t = useTranslations("AppHeader");
